Don't strip the OAuth response!
Summary:
I guess we were assuming the response is always json, and thus
safe/desirable to strip?  But in my case the response was
zlib-compressed, and stripping off newlines caused the zlib
decompression to fail.  I think it's better not to try to munge with
the output in this way.  If callers really care, *they* can strip...

Reviewers: benkomalo, jace

Reviewed By: benkomalo

Differential Revision: http://phabricator.khanacademy.org/D24