Add logic to look for logs in multiple log-versions if necessary.
Summary:
This code deals with the case where we deploy and flip the switch to
change the default app-version, and later try to fetch logs from
before the switch-flip.  By default, such a fetch will return no
results, because unless you tell it otherwise, appengine only returns
logs associated with the currently active app-version.  Looking at the
previous fetches will tell us what version was current when the
previous logs were generated, and we can tell that to appengine.

This makes use of functionality from a previous commit, that makes
fetch-logs emit what app-version is current, and logs that data in the
-error.log files.  This commit looks in previous -error.log files to
find the app-versions mentioned there.  The sh script passes this info
to the .py script, which uses those app-versions if fetching with the
default app-version returns no logs.

Ugh!, complicated.  There are simpler ways to have done this, but they
lose the ability to easily run fetch_logs.sh against logs from a few
days ago.  With this scheme, we log all the data we need for
reproducibility, and use that data automatically.

Test Plan:
Ran manually.  Will need to do more testing once the changes to
/api/v1/fetch_logs have been deployed (should be on 18 May 2012 or 21
May 2012), and we actually start logging the header functionality.

Reviewers: benkomalo, jace

Reviewed By: jace

Differential Revision: http://phabricator.khanacademy.org/D36