Fix the date-parsing routines.  Also give a bit more info when logging.
Summary:
date is a bit finicky: even though I want it to parse
   2012-05-18 15:00:00 -1 hour
to mean
   2012-05-18 14:00:00
it actually parses it to mean
   2012-05-18 10:00:00 PDT 2012
(I think that it is intepreting the -1 as a timezone specification of
some sort, though I'm not sure.)

Luckily, I can get around this by doing
   2012-05-18 15:00:00 last hour
which is what I do.

Somewhat unrelatedly, I changed fetch_logs.py to emit how many
loglines it's read, just to provide a nice summary.

When doing that, I realized the error.log is increasingly misnamed --
it has error information, to be sure, but also all sorts of other
interesting information about how the log-fetching has been going.
I've renamed it to status.log.

Reviewers: jace, benkomalo

Reviewed By: jace

Differential Revision: http://phabricator.khanacademy.org/D46