Fix the try-a-different-version code, and do a bit more error-logging.
Summary:
The code that tried to detect if fetching logs at a particular version
was failing, was buggy because when it tried to discern whether the
log output was empty, it was looking at the compressed string.  This
will never succeed because the compressed version of the empty string
is like 16 bytes long.

Now I do the decompressing in fetch_logs, so I can compare-to-empty on
the decompressed output.

To help figure out what's going on with versions, I added some print
statements.  And I added a robustness check to avoid an IndexError.

Reviewers: benkomalo, jace

Reviewed By: benkomalo

Differential Revision: http://phabricator.khanacademy.org/D58