Allow overriding archive dir from cmd line args
Summary:
Let's move towards a place where config params can be specified on the
command line instead of hardcoded as paths in config files. The config
files will be a nice place to house defaults, and production setup. But
it'll be easier for unit testing and development if these can be
overridden or specified on the command line.

The alternative is to keep a separate config file just for development,
and add it to .gitignore, but that seemed heavier.

Test Plan: Ran on my devbox manually while specifying the archive_dir

Reviewers: yunfangjuan

Reviewed By: yunfangjuan

Differential Revision: http://phabricator.khanacademy.org/D56