A very crude topic tree fetcher.
Summary:
This script analyzes TopicVersion entities stored in mongo, and fetches
the actual topic tree content for each version (that we don't already
have info for), saving it in the same mongo db.

Note that the topic trees don't correspond to a single entity, so the
existing backup pipeline doesn't save that data. It is quite large, but
changes relatively infrequently, so it's fine to save it all for
convenience, I think.

A TODO later is to export this from the mongo db to hive if we want it
used there, but I'm not sure how to do that yet.

Reviewers: jace, yunfangjuan

Reviewed By: jace

Differential Revision: http://phabricator.khanacademy.org/D188