Add an --ignore-before flag, which lets us ignore a prefix of log records.
Summary:
The intended use is to use ignore_before='"_gae_bingo_cache" (fail)'
as a way to avoid records before a memcache flush.  This is because
_gae_bingo_cache should always be in the memcache under normal
circumstances -- it's looked up all the time, and memcache is an LRU
cache -- so if we see a failure, it must be due to a memcache flush.
If we start processing at the "(fail)", we're processing right at
memcache-flush time.

Fix --ignore-before to mean 'before in time', not 'before in the logs'.

The logs, sadly, do not necessarily come sorted in chronological
order.  The semantics of --ignore-before are supposed to be
chronological, not just how the logs happen to have ended up in the
logfile.  This change makes that so.

Add a warning message if the ignore-text is never found.

Add a note about how many requests we ignore due to ignore_before.

This is just informational.

Even more information in the informational message!

Change the example ignore_before text for finding the memcache flush.

It turns out, on analysis, that
   '__layer_cache_setting_model._get_settings_dict__" (fail)'
is a more reliable test of flushing than _gae_bingo.  _gae_bingo
*does* get evicted during the normal course of operations, at least
occassionally, but I never say _get_settings_dict__ fail in the hour
of logs that I looked at (except at cache-flush time).

Reviewers: kamens, jace

Reviewed By: kamens

Differential Revision: http://phabricator.khanacademy.org/D212