A crude script to do a crude import to mongo.
Summary:
This reads raw Hive table data from S3 to import to mongo. There were a
couple of options to get this data to the reporting db and they all kind
of sucked: in the end it seemed easiest to just do the simple thing.
This is a really simple and raw solution.

The summary/reporting tables should probably have date as a column being
summarized, but I'll talk more about that in a later change.

This also introduces boto, which I think is probably going to be the
most flexible AWS tool.

Reviewers: david, yunfangjuan, jace

Reviewed By: david

Differential Revision: http://phabricator.khanacademy.org/D210