Add a ParsedMemcacheLine class and use it to access memcache info.
Summary:
No more memcache_line[1] and memcache_line[2].startswith('get').  Now
we can do 'memcache_line.key' and 'memcache_line.does_get()'.

The biggest semantic change is that incr is now treated the same as
get.  This changes some of the stats somewhat, particularly ones
involving _gae_alternative (the only key that we use incr with).

Treat incr more like a set command than a get command.

The 'get before set' stats were way off because they were treating
incr as a 'get', even though it acts like a 'set'.  I've fixed the
code to separate out what the logs think of as get-like and set-like,
from what the semantics determine are get-like and set-like (probably
the logs should be fixed up to match the semantics as well...)

Add comment on why we store memcache lines intermediate-ly as tuples.

Add a few accessor methods to make ParsedMemcacheLine even clearer.

We now call a function rather than explicitly comparing success
against True and False.  This adds clarity because success can also be
None, which is non-obvious.  The access methods hide that bit.

Test Plan:
Did a diff against the output of this script on a log-hour
both before and after this change.

Reviewers: jace, kamens

Reviewed By: kamens

Differential Revision: http://phabricator.khanacademy.org/D228