Have memcache_lines() sort all memcache operations by time, not just per-req.
Summary:
Before, if two requests were happening at the same time, and their
memcache operations were interleaved, we'd return all the memcache ops
from request 1, then all the ones from request 2.  But this doesn't
accurately reflect how things work: suppose we have:

REQ 1 -- start-time t   REQ 2 -- start-time t-1
                        lookup key y -- success
lookup key x -- fail
set key x
                        lookup key x --  success

In this case, req 2 comes before req 1, so with the old code, we'd see
the successful lookup of x (in req 2) before the 'set' of x (in req
1).  This will confuse our analysis.

I fixed this by storing all the memcache operations in a global list,
which we sort before iterating.  That way we'll just see

   lookup key y -- success
   lookup key x -- fail
   set key x
   lookup key x --  success

and everything is in the correct order.

This makes the code a bit messier than it used to be, because we still
store the per-request memcache lines as well.  We don't use that data
very much, so I could be more efficient here with a little work, but I
wanted to leave it around to ease future analysis.

Reviewers: kamens, jace

Reviewed By: jace

Differential Revision: http://phabricator.khanacademy.org/D240