Change the way video metrics tables are created.
Summary:
This paves the way for a more permanent set of tables, as we can now
specify prefixes for the location (so we can specify the final,
permanent location for use with the report). This also partitions the
results by date ranges.

I plan to have this run daily, weekly, and monthly

Test Plan:
    none, et.

Reviewers: yunfangjuan, david, jace

Reviewed By: jace

Differential Revision: http://phabricator.khanacademy.org/D230