Print the time-range for get_but_never_set.
Summary:
After this analysis, which shows all the get-but-not-set's happen
within 25 seconds of each other, on an hour-long log file -- it's
clear that the flush_memcache operation isn't instantaneous (probably
the command is sent asynchronously to all the memcache shards?) so
even after the flush, some 'gets' might succeed for a few seconds sine
the clear command hadn't gotten to the relevant shard yet.

Add another key prefix.

Reviewers: jace, kamens

Reviewed By: kamens

Differential Revision: http://phabricator.khanacademy.org/D241