Introduce a deploy script.
Summary:
This does little more than the s3cmd put that was in the Makefile
But it does notify hipchat and also allows you to do more precise
filtering, and also lets you deploy to dev subdirectories later (if
needed). It also allows us to do version tracking on prod, which I'd
like to do at some point.

Reviewers: david, jace, yunfangjuan

Reviewed By: jace

Differential Revision: http://phabricator.khanacademy.org/D251