Fix SOURCE directives to use absolute paths.
Summary:
When using ADD FILE from s3 URL's, the file gets downloaded to a local
download directory. SOURCE commands in Hive seem to need an absolute
path to reference those.

This works until the hive version changes (which there is a TODO to
fix), since the path includes the hive version (though is
otherwise stable for AWS' hadoop setup)

Reviewers: yunfangjuan, david

Reviewed By: david

Differential Revision: http://phabricator.khanacademy.org/D254