code to create a training dataset for topic/exercise Bayes nets
Summary:
To learn the Bayes nets, I need a user-by-exercise matrix
for each topic.  Each element in the matrix should represent
something like a primitive accuracy estimate for the user in
that exercise.  As a placeholder, right now I use an EWMA of
accuracy, but this is something I will experiement with.

Test Plan: I have run this on small sets of ProblemLogs

Reviewers: yunfangjuan, benkomalo, josh

Reviewed By: benkomalo

CC: david

Differential Revision: http://phabricator.khanacademy.org/D257