segment by topic for accuracy_deltas table in hive
Summary:
It's a little meaningless to just see incremental gains of students who
have done exactly X stacks across all topics, since there are topics that can be
completed in 2 stacks and topics which have more than 16 skills.

Also, the final table to which the reducer writes is now partitioned as
a labelling mechanism, as @benkomalo did for some video summary tables. This
eliminates the need to pass in a separate suffix argument to label the table
with the start and end dates which are already passed in.

Reviewers: tom, benkomalo

Reviewed By: benkomalo

CC: jace, yunfangjuan

Differential Revision: http://phabricator.khanacademy.org/D260