Disable computing the 'total' row for topicsummary
Summary:
This is basically just the total user count for the date range, which is
somewhat duplicated in the user_video_distribution table anyways, and is
easily aggregated after the fact in the reporting tools, since it's just
a sum over a small number of topics.

Not really sure of a good solution otherwise.

Reviewers: yunfangjuan, jace

Reviewed By: jace

CC: david

Differential Revision: http://phabricator.khanacademy.org/D263