Add in a script to query Hive metadata.
Summary:
This uses SSH to run MySQL commands on a Hive master node.
Opening up ports for the MySQL daemon is more dangerous and is annoying
since the mysqld runs as root.

We all have SSH credentials to access this: I'm just going to keep it
simple and make this script leverage that and use ssh to run the mysqld
commands.

Reviewers: jace, yunfangjuan

Reviewed By: jace

CC: david

Differential Revision: http://phabricator.khanacademy.org/D268