save a pseudo-topic "any" for aggregating all topics on hive
Summary:
We now segment data by topics. If we want to view stats across all topics, we
have to aggregate topics in mongo or on the client-side JS. The disadvantage of
doing it client-side is that we'd have to fetch rows for all topics, increasing
the amount of data we need to get by 30 times or so. The disadvantage of doing
this in pymongo on the flask app is that we'd be duplicating logic on the client
side, since the client side is currently responsible for forming the query to
fetch rows from mongo. (I don't trust group by queries via sleepy mongoose:
https://jira.mongodb.org/browse/SERVER-5874)

This just does the aggregation as a natural part of the Hive map reduce and
saves us computation further on, and is also a simple solution.

Reviewers: jace, benkomalo

Reviewed By: jace

CC: yunfangjuan, tom

Differential Revision: http://phabricator.khanacademy.org/D270