update UI with batch requests, and tries to be user-friendly
Summary:
See, I have a very slow internet at home so this sort of stuff makes
quite a difference for me! It takes about 30 seconds to load all at once the
entire data for any number of stacks for any topic, and batched up it's 35
seconds but I get UI updates every 2-3 seconds or so. World of difference for
the user. I guess at the office this will be less of an issue...

Man, I am more productive in 1 hour at 2 am at home on the weekend than an
entire day at the office in the weekday, lol. I should just "call in sick" and
work from home some more lol (and maybe turn off hipchat and email or something).

Test Plan: Banged on the UI, tried changing options in the middle of a request

Reviewers: benkomalo

Reviewed By: benkomalo

CC: jace, yunfangjuan, tom, jlfwong

Differential Revision: http://phabricator.khanacademy.org/D285