add support for adding multiple series to the graph
Summary:
Not sure if this is the final form it will take, but at least it's fun to play
with for now. In the future, once I actually get segments to work, maybe we only
want to overlay different segments but keep # stacks done, which topic(s)
constant?

Test Plan: ran it

Reviewers: benkomalo, jace, tom

Reviewed By: tom

CC: tom, yunfangjuan

Differential Revision: http://phabricator.khanacademy.org/D291