Exporter script for video matrix to GAE
Summary:
The actual POSTing is stubbed out in place of a dev call that
essentially does a simple READ (to verify access works). But the reading
is tested with an existing table I generated and works with the pruned
data.

We may want to send it up in batch, but it's not terribly big. If
compressed (as in use numeric indices instead of the large GAE key
string values, not gzip), it's probably about 1.5Mb

Test Plan: Ran it against s3://ka-mapreduce/tmp/video_cooccurrence_20120501_20120620 without any errors

Reviewers: jace, josh

Reviewed By: jace

CC: yunfangjuan

Differential Revision: http://phabricator.khanacademy.org/D301