Add in support for deploying to a "branch"
Summary:
Branches will go to dirnames of the form
s3://ka-mapreduce/code/<branchname>/

By convention, the top level directory is the main or production branch,
so keep that clean. There is also a 'dev/' branch we've been using, but
if you want your own, just name it 'branch-<yourname>' by convention, so
the script can ignore it when reading files.

Test Plan: ran it manually

Reviewers: jace

Reviewed By: jace

CC: david, yunfangjuan

Differential Revision: http://phabricator.khanacademy.org/D330