Update hostname of reporting db.
Summary:
This does not make any attempt to fix the duplication issue that's
apparent in the codebase: it's just a simple dummy search/replace to
point mongo requests to the new location of the reports db.

Test Plan: Ran a local dashboard, and checked that requests to sleepymongoose worked fine.

Reviewers: david, jace

Reviewed By: jace

CC: yunfangjuan

Differential Revision: http://phabricator.khanacademy.org/D348