Fix a bug in user_video_summary.
Summary:
This query wasn't compiling anymore since the MAX/MIN functions in Hive
are used to find the max/min values in a column, and not between two
passed in arguments. Instead, change it so that it uses a more awkward
IF statement (not sure how to do this otherwise).

Also change the type of "completed" to a BOOLEAN, as it seemed more
natural. This makes it slightly more awkward to count the number of
completed videos, but I felt it to be a more natural representation of
the data.

Test Plan:
Ran it for 2012-06-08 on interactive Hive cluster and sanity checked the data.

If you guys are OK with this change, I plan to run it for every day in May1 to today on the interactive Hive cluster

Reviewers: yunfangjuan, david

Reviewed By: david

Differential Revision: http://phabricator.khanacademy.org/D363