Support for downloading _GAEBingoRecordIdentity
Summary:
This adds gae_bingo (https://github.com/Khan/gae_bingo) as a "fake
submodule"). (explained in
http://debuggable.com/posts/git-fake-submodules:4b563ee4-f3cc-4061-967e-0e48cbdd56cb)

The bingo codebase is needed since the existing entity downloading
framework works with somewhat raw protobufs of db.Model objects. The
protobuf of the bingo models we wish to download have pickled data,
which we need to unpickle to get in a JSON format (which is ultimately
what we need to store in Hive)

Test Plan:
1. Created a cfg/test.json that tries to download _GAEBingoIdentityRecord.
2. Created src/oauth_util/dev_consts.py so that OAuth fetch requests would go against my dev server
3. Ran src/gae_download.py (with some lines commented out that talks to mongo)
4. Verified that the download emited JSON files in a local directory

Reviewers: kamens, jace

Reviewed By: jace

CC: yunfangjuan

Differential Revision: http://phabricator.khanacademy.org/D572