Add in table schemas for GAEBingo entities.
Summary:
This involves creating incremental tables and snapshot partitions for
GAEBingoIdentityRecord, identical to the process used for UserData
(since these update in a similar to fashion to UserData, and in
 practice there will be a 1:1 correlation for these entities to
 UserData for active users).

GAEBingoAlternativ entities are simpler and the entire set is going to
be downloaded in entirety periodically, as it's pretty small.

Test Plan: Ran the table creation queries on Hive interactive to make sure they're well formed. Have not actually tested the full import process yet.

Reviewers: kamens, yunfangjuan

Reviewed By: kamens

CC: jace

Differential Revision: http://phabricator.khanacademy.org/D624