Hook up bingo_download scripts to daily scripts.
Summary:
This completes the hookup into the daily cron (since it calls
load_emr_daily.sh) and does the bingo download and uploads the data
to Hive for the alternatives data.

Also fix the name of the table in bingo_download.py and fix the file
permissions

(Still have to do similar work for the BingoIdentityRecord)

Test Plan: None yet. These scripts don't make it easy to test locally, unfortunately, so I will test manually on analytics once the pieces are in place.

Reviewers: david

Reviewed By: david

CC: yunfangjuan

Differential Revision: http://phabricator.khanacademy.org/D635