Add in clarifying comments about bulk_download
Summary:
Also fix an issue where PYTHONPATH was not set properly. This wasn't an
issue on the analytics machine because there is some magic in the
bashrc that puts /usr/local/google_appengine on the PYTHONPATH. Not
relying on that is nice (also makes it easier to test locally)

Test Plan: Ran src/bulk_download.py and src/gae_download.py locally to ensure it doesn't explode on syntax issues.

Reviewers: jace, yunfangjuan

Reviewed By: jace

Differential Revision: http://phabricator.khanacademy.org/D673