Fix table location of bingo_alternative_info
Summary:
This is once of those snapshotted "factual tables", and it's defined
with the P suffix (for partition) in ka_hive_init.q

Test Plan:
None for directly testing this.

Running s3 commands to move the files from the incorrect location (Running s3 commands to move the files from the incorrect location (../bingo_alternative_info/...) to the new, correct one (.../bingo_alternative_infoP/...) and then run ka_hive_init.q on the interactive Hive cluster and see if I can query the table.

Reviewers: jace

Reviewed By: jace

Differential Revision: http://phabricator.khanacademy.org/D683