Create UserData/GAEBingoIdentityRecord sample views.
Summary:
The VIEW's screw up with the bucketing, and it's often useful to just
use a small bucket for testing queries. This doesn't solve the general
problem of the lack of bucketing now, but it does help with the testing
for now.

Test Plan: Ran on Hive interactive to make sure nothing blew up.

Reviewers: yunfangjuan

Reviewed By: yunfangjuan

CC: jace

Differential Revision: http://phabricator.khanacademy.org/D684