Add Hive script/reducer for user experiment info.
Summary:
This generates a summary of all users participating in an A/B
experiment, and partitions them by alternative, so that developers can
then do additional queries for users only within a partition.

Test Plan:
Has not been tested in full.
I ran the reducer locally with some faked input to ensure it doesn't explode, and ran part of the subquery on a Hive interactive to see if it's well formed.

yet to fully verify correctness (need to get live data in the Hive tables for the experiment data first) - want to get initial version reviewed to see if overall approach makes sense

Reviewers: david, jascha, jace

Reviewed By: jace

CC: yunfangjuan, jace

Differential Revision: http://phabricator.khanacademy.org/D647