add some utility modules for sending notifications by email and HipChat, and use them for the major daily jobflows
Summary:
Beyond the commit message, it should be self explanatory with
one exception:  I did move the hipchat.cfg to the same directory as the
module (notify.py) that uses it.  You will need to do this in your
local repo, so that deploy.py still works for you.

Test Plan:
I tested small pieces manually/locally, but at this point further
testing will be done best by releasing to the analytics machine (after review).
It would be great to do that soon, as I am OOO Thurs/Fri.

Reviewers: benkomalo, yunfangjuan

Reviewed By: benkomalo

Differential Revision: http://phabricator.khanacademy.org/D718