Install the Python library lxml
Summary:
The aws-config scripts for analytics rely on requirements.txt to
determine which Python libraries are installed.

lxml is needed for src/gae_dashboard/instance_report.py

GZIP stats JSON

It's text data that is much smaller gzipped.

Since the resulting file has the GZIP header, it won't be empty and we
can't trivially tell if it should be deleted. You will be able to which
files are invalid because they will be about 25 bytes on disk.

Test Plan:
Installed lxml manually on the analytics machine and ran
src/gae_dashboard/fetch_logs.sh which invokes the instance report.

Fingers crossed.

Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: http://phabricator.khanacademy.org/D801