Wrap daily_request_log_urlroute_stats cursor in a generator
Summary:
The problem was that none of the daily_request_log_urlroute_stats
results were being displayed on the page. The loop that set the "url"
field to the value of the "urlroute" field was iterating through all
results and exhausting the cursor.

Instead, modify results one by one as they are returned by the cursor.

Test Plan: Verified that page fetches data properly during local development.

Reviewers: benkomalo

Reviewed By: benkomalo

Differential Revision: http://phabricator.khanacademy.org/D846