Change the script to do all web tests in parallel.
Summary:
Before this change, all urls from a single location were done in
parallel, but different locations were done serially.  No reason for
that.  But fixing it required separating out the batch run into two
steps, one where we submit the urls (different options for different
locations), and one where we get the results of the test (same for
everyone).

Test Plan:
To run it manually.  But we hit our quota for using webpagetest.org
today, so it's crossed fingers until tomorrow...

Reviewers: chris

Differential Revision: http://phabricator.khanacademy.org/D871