Change the script to do all web tests in parallel.
Summary:
Before this change, all urls from a single location were done in
parallel, but different locations were done serially.  No reason for
that.  But fixing it required separating out the batch run into two
steps, one where we submit the urls (different options for different
locations), and one where we get the results of the test (same for
everyone).

Fix a typo: I was meaning to call optparse with argv, but forgot to.

Add support for 'test mode' where we only fetch 2 urls.

This is to keep us from exceeding the day's quota when testing the
script.

I supported this using commandline flags, so I added one for verbosity
too, and refactored the file a bit to allow for more flags later.

Default test-mode to true for right now.  I keep forgetting to specify -t.

Test Plan:
To run it manually.  But we hit our quota for using webpagetest.org
today, so it's crossed fingers until tomorrow...

Reviewers: chris

Reviewed By: chris

Differential Revision: http://phabricator.khanacademy.org/D871