Fix the code to have 4-space indents.
Summary:
I am reluctant to make such lint-only changes, since it will make it
more difficult to merge in changes to these files in the future, but
I'm already hacking on these files enough that changes are going to be
difficult to incorporate.  And I can't take all the phabricator
complaining anymore.

Get rid of an unused import.

Test Plan: Ran run_webpagetest.py and it worked like before.

Reviewers: chris

Reviewed By: chris

Differential Revision: http://phabricator.khanacademy.org/D873