Finish the tooling to store webpagetest data in mongo-db.
Summary:
Add a gitignore file for api_key.

api_key is a 'secrets' file, so it's not checked into git.

Add in a missing import.

Didn't notice this in testing because it only matters for an error
situation that we don't have.

Test Plan:
Ran the script manually, and verified it said it saved dicts to
mongo.  I haven't actually looked at mongo to make sure the data is
there, though (I don't know how).

Reviewers: chris

Reviewed By: chris

Differential Revision: http://phabricator.khanacademy.org/D876