When unable to get data from yesterday, try the day before.
Summary:
It can take a while before yesterday's data is fully processed.
Before, while waiting for that the dashboard would return 0 results.
Now it goes to the day before yesterday instead.  In fact, it's
willing to go back a week.  This handles the situation that some days
were skipped entirely due to errors of some sort.

I now say what day we're showing results for, as part of the the html
page.  If the day is a long time ago, I print a message to that
effect.

Quote attribute values.

Test Plan:
Ran './main.py -d' to test locally, and visited the following urls to
make sure they showed results:
   http://localhost:5000/gae_stats/daily_request_log_url_stats
   http://localhost:5000/gae_stats/daily_request_log_urlroute_stats
   http://localhost:5000/gae_stats/url_stats
They did, and had '2012-10-29' in the <h2> (when appropriate).

I then changed main.py's utc_as_dt() to replace 'utcnow()' with
'(2012, 11, 1)' -- that is, pretending it's two days from now.  I then
visited the pages again, and showed that they still said '2012-10-29'
in the <h2>, but the <h2> also said 'OLD DATE'.

I then changed utc_as_dt() to pretend today is (2102, 10, 31), and
visited the pages again.  They still said '2012-10-29', but did not
say 'OLD DATE', since two days ago is not considered old.

Reviewers: chris, alpert

Reviewed By: alpert

Differential Revision: http://phabricator.khanacademy.org/D885