The referer can be '-' to indicate no referer.
Summary:
This happens when someone just types a url into an address bar.  It
frequently happens on the khan academy homepage, which is why we
were severely undercounting the homepage in the logs before (by a
factor of 4!) -- we were skipping all loglines that had '-' as a
referer.

Now when the referer is missing ('-'), we store it as the empty string
in the parsed logs.

Test Plan: Manually tested on a url with a '-' referer.

Reviewers: alpert, jace

Reviewed By: jace

Differential Revision: http://phabricator.khanacademy.org/D888