fix race condition on daily reporting
Summary:
Daily ex stats dashboard has been screwing because it was possible for it to
kick off before the new UserData/Bingo/Scratchpad/Feedback paritions were created.
This also fixes a bug with user_feedback_summary - it needed to be moved somewhere
whence it would run only after the new Feedback partition was actually there.

Test Plan:
none.  :( I have no good test environment set up for changes like this,
which is a bit of a problem. On the other hand, no users are affected if if fails...
only me.

Reviewers: benkomalo

Reviewed By: benkomalo

Differential Revision: http://phabricator.khanacademy.org/D894