fix the was user_feedback_summary is computed
Summary:
user_feedback_summary has always been doing the wrong thing (my bad!),
in that if a Feedback entity was updated, it considered that an activity
for that author of that entity.  But most updates are votes/replies/etc,
which should not be considered activity of the authtor.  This reduces the
number of user actions by a factor of approximately 4 or 5.

Test Plan: tested manually

Reviewers: benkomalo

Reviewed By: benkomalo

Differential Revision: http://phabricator.khanacademy.org/D896