Take the median value of each run, rather than the mean.
Summary:
This is an attempt to smooth out the data some.  We're seeing very
noisy data, probably just due to the fact we have very few data points
(3 for each query).  I'm hoping that taking the median rather than the
mean will help smooth things out some.  It may not make much of a
difference.

Reviewers: alpert

Reviewed By: alpert

Differential Revision: http://phabricator.khanacademy.org/D931