Graph historical instance count, qps, latency, memory usage
Summary:
This data has been available in the instance summary but was until now
not surfaced. Most work was on the frontend to build a field selector.

Load time is slow because there are many data points. Speed up is left
for future work if it proves useful.

Test Plan: Tested locally with main.py -d against live data.

Reviewers: benkomalo

Reviewed By: benkomalo

Differential Revision: http://phabricator.khanacademy.org/D967