fix errors with video_topic - some videos no longer have topic_string_keys
Summary:
The change in https://khanacademy.kilnhg.com/Code/Website/Group/stable/History/d38a55ea7d3d
broke the topic mappings used for the video usage dashboards. Enitites used to store the
the topics they were mapped to, but now it is an @property that is not available within
out data pipeline.

 I need to talk to Tom, but I'm guessing
he'll want to keep this change and we'll need to come up with a new, suitable way to map
all of these things to topics.  :(  For the time being, I'll check in this change to stop the errors.

Test Plan: tested locally

Reviewers: csilvers

Reviewed By: csilvers

CC: tom

Differential Revision: http://phabricator.khanacademy.org/D987