Scrape data from GAE's main dashboard into mongo db
Summary:
parsers.py now accepts /dashboard markup and can extract the URLs
corresponding to the charts at the top of the page for 6, 12, and 24
hour charts.

The new dashboard_report.py can unpack the data in a Google Charts API
URL and turn it back into time series data for each of the charts.

Next up is aggregating the time series data and saving records that are
newer than what is in the database, then, fetch_stats.sh needs to be
updated to run the full pipeline.

Test Plan:
I ran:
<~/private_pw ./gae_dashboard_curl.py "https://appengine.google.com/dashboard?app_id=s~khan-academy" USERNAME >dashboard.html

to get a local copy of the dashboard, then tested output of:
<dashboard.html dashboard_report.py `date +%s`

Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: http://phabricator.khanacademy.org/D998